Copy over the lock file, as well (#6)
This commit moves over the lock file in the Dockerfile command. It was
previously commented out because of some tomfoolery when doing local
development.

I was trying to install Gems in Docker but didn't have a `Gemfile.lock`
file yet, so I removed the line that copied the lock file to generate a
`Gemfile.lock`. I never changed it back :(