Add users (#9)
* Fix up cop about logging instead of puts

* Add User model using Lockbox + Blind Index

This commit adds the user model to the application. Lockbox is used to
encrypt attributes for the user, and we use Blind Index to query the
e-mail field.

* Configure Lockbox

* Add default user spec

* Update migration to encrypt more attrs

This commit changes the user schema to encrypt more attributes.

We will be only using blind indexing for first name, last name and
email.

Email should use a blind index because it is the identifier that the
user will use to log in. Without this, the login flow would be a pain.

First and Last name both will leverage a blind index so that we can
query data brokers for exact matches. If we needed something like
figuring out the correct age + state, we can further drill down in
memory.

* Specs for user encrypted attrs

* Auto loading warning fix